Add a route for requesting images.
This route accepts the parameters id and size. Valid parameters
for size can be one of: smallest, small, medium, large. This calls
the specified Image filter which first checks for a cached version
of the image, if not found then fetches the source image, invokes
the filter class, the filter resizes the source image, saves a copy
in the cache and returns the image response.

Images are cached for 30 days at present.